refactors and formatting

+++ include/c938_metal/c938_text.h
@@ -17,7 +17,7 @@
 
 [[fragment]] metal::float4 c938_text_fragment(
   struct c938_data_vertex_textured_coloured,
-  metal::texture2d<half>
+  metal::texture2d<float>
 );
 
 #endif
